feat: send also configuration with rendered event
So we can subtract `asyncRenderResolvedTimeout`
from performance metrics.

risk: low
JIRA: STL-1064
